pyproject.toml
[tool.poetry]
name = "weasley-clock"
version = "0.1.0"
description = "GUI app for monitoring network devices and showing a picture when selected devices are connected"
authors = ["Joshua Crim <josh@jshcrm.com>"]

[tool.poetry.dependencies]
python = "^3.7.6"
kivy = "^1.11.1"
pillow = "^7.2.0"
python3-nmap = "^1.4.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
